add/utilize latest auto-update library
