+++ app.c
@@ -1,5 +1,6 @@
 #include <alloc/alloc.h>
 #include <flip_storage/storage.h>
+#include <update/update.h>
 
 // Entry point for the FlipWorld application
 int32_t flip_world_main(void *p)
@@ -55,7 +56,7 @@ int32_t flip_world_main(void *p)
     save_char("app_version", VERSION);
 
     // for now use the catalog API until I implement caching on the server
-    if (flip_world_handle_app_update(fhttp, true))
+    if (update_is_ready(fhttp, true))
     {
         easy_flipper_dialog("Update Status", "Complete.\nRestart your Flipper Zero.");
     }
