+++ update/update.h
@@ -1,0 +1,8 @@
+#pragma once
+#include <flip_world.h>
+#define BUILD_ID "676900d983aa88302bc114c6"
+#define APP_ID "flip_world"
+#define FAP_ID "flip_world"
+#define APP_FOLDER "GPIO"
+#define MOM_FOLDER "FlipperHTTP"
+bool update_is_ready(FlipperHTTP *fhttp, bool use_flipper_api);
